pyproject.toml
[project]
name = "sales-db"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apache-airflow>=3.1.8",
    "boto3>=1.42.73",
    "dbt-core>=1.11.7",
    "dbt-snowflake>=1.11.3",
    "dotenv>=0.9.9",
    "faker>=40.11.0",
    "nba-api>=1.11.4",
    "pandas>=3.0.1",
    "requests>=2.32.5",
]
